optimize

+++ api/core/rag/datasource/vdb/tencent/tencent_vector.py
@@ -64,9 +64,22 @@ class TencentVector(BaseVector):
             "vector_store": {"class_prefix": self._collection_name}
         }
 
+    def _has_collection(self) -> bool:
+        collections = self._db.list_collections()
+        for collection in collections:
+            if collection.collection_name == self._collection_name:
+                return True
+        return False
+
     def _create_collection(self, dimension: int) -> None:
         lock_name = 'vector_indexing_lock_{}'.format(self._collection_name)
         with redis_client.lock(lock_name, timeout=20):
+            collection_exist_cache_key = 'vector_indexing_{}'.format(self._collection_name)
+            if redis_client.get(collection_exist_cache_key):
+                return
+            if self._has_collection():
+                return
+
             self.delete()
             index_type = None
             for k, v in enum.IndexType.__members__.items():
@@ -107,6 +120,7 @@ class TencentVector(BaseVector):
                 description="Collection for Dify",
                 index=index,
             )
+            redis_client.set(collection_exist_cache_key, 1, ex=3600)
 
     def create(self, texts: list[Document], embeddings: list[list[float]], **kwargs):
         self._create_collection(len(embeddings[0]))
